Moved the session_* functions. session_* functions should be able to deal with a SessionDB instance.
